add the certbot layer

+++ src/domain.py
@@ -7,16 +7,8 @@
 # if will redirect anything to home , and the new created subdomain will be forwarded to the right subdomain , the attack can filter the html string to actualy detect default redirection
 
 ###################### THIS IS PROTOTYPE NOT PRODUCTION READY ######################## 
-
-
-# missing seralization of
-
-
-
 # the usage of this will be docker run -it -p 80:80 -p 443:443 -v /<path>/:/etc/nginx/conf.d , where conf.d contain the config file inside
 # <path>
-
-
 # GODADDY
 # 60 request per minute , 429 and an error message indicating that your rate limit has been surpassed 
 
@@ -28,19 +20,23 @@ import secrets
 import os
 import subprocess
 import requests
+import base64
+
 
 
 # used to make the microservice more dynamic
-path_to_template = "/etc/nginx/conf.d/template.sample"   # the sample must exist in the docker volume 
+path_to_template_auto = "/etc/nginx/conf.d/templatesecauto.sample"   # the sample must exist in the docker volume 
+path_to_template_manu = "/etc/nginx/conf.d/templatesecmanual.sample"
 config_output_path = "/etc/nginx/conf.d"
-DOMAINNAME= os.environ.get('DOMAINNAME')
-API_KEY = os.environ.get('API_KEY')
-API_SECRET = os.environ.get('API_SECRET')
+DOMAINNAME= os.environ.get('PY_DOMAINNAME') # domain name must be complete with *.com
+API_KEY = os.environ.get('PY_API_KEY')
+API_SECRET = os.environ.get('PY_API_SECRET')
 URL= "https://api.godaddy.com"
 HEADERS = { 
     "Authorization" : f"sso-key {API_KEY}:{API_SECRET}",
     "accept" : "application/json" 
     }
+HOST= os.environ.get('PY_HOST')
 
 app = Flask(__name__)
 
@@ -73,47 +69,53 @@ def checkdomaingodaddy(domain):
 
 
 
-def adddomaindaddy(host,domain):
-    # add the new domain to the go daddy first
+def adddomaindaddy(domain):
     headers = { 
-    "Authorization" : f"sso-key {API_KEY}:{API_SECRET}",
-    "accept" : "application/json",
-    "Content-Type" : "application/json" 
+        "Authorization": f"sso-key {API_KEY}:{API_SECRET}",
+        "accept": "application/json",
+        "Content-Type": "application/json" 
     }
 
     data = [{     
-        "data": host,
+        "data": HOST,
         "name": domain,
         "ttl": 600,
         "type": "A"
     }]
 
-    response = requests.patch(URL + f"/v1/domains/{DOMAINNAME}/records",headers=headers,json=data)
+    response = requests.patch(URL + f"/v1/domains/{DOMAINNAME}/records", headers=headers, json=data)
 
-    if ( response.status_code == "200" ):
-        return 0
-    elif ( response.status_code == "429"):
-        print("stored in the memory and served")
-        return 2
-    elif ( response.status_code == "404"):
-        print("domain not found")
-        return 1
+    if response.status_code == 200:
+        return None  # No error, return None
+    elif response.status_code == 429:
+        return "Rate limit of GoDaddy API triggered. Please handle this asynchronously."
     else:
-        return 3
+        try:
+            error_message = response.json().get('message')
+            if not error_message:
+                error_message = "An unknown error occurred with GoDaddy API."
+            return error_message
+        except Exception:
+            return "An unknown error occurred with GoDaddy API."
     
 def removedomaindaddy(domain):
     # remove domain 
-    response = requests.delete(URL + f"/v1/domains/{DOMAINNAME}/records/A/{domain}",headers=HEADERS)
-    if response.status_code == "204":
-        return 0
-    if response.status_code == "404":
-        print("record name not found")
-        return 1
-    if response.status_code == "429":
-        print("stored in the memory and servred")
-        return 2
+    response = requests.delete(URL + f"/v1/domains/{DOMAINNAME}/records/A/{domain}", headers=HEADERS)
+    
+    if response.status_code == 204:
+        return None  # No error, return None
+    elif response.status_code == 404:
+        return "Record name not found."
+    elif response.status_code == 429:
+        return "Rate limit of GoDaddy API triggered. Please handle this asynchronously."
     else:
-        return 3
+        try:
+            error_message = response.json().get('message')
+            if not error_message:
+                error_message = "An unknown error occurred with GoDaddy API."
+            return error_message
+        except Exception:
+            return "An unknown error occurred with GoDaddy API."
 
 # Endpoint 1: create a random subdomain for a user and nginx config file generated ( this require no mapping or tracking thus it used to generate the file and subdomain , storing and tracking handled at another layer )
 @app.route('/createrandomdomain', methods=['GET'])
@@ -126,26 +128,23 @@ def create_random_domain():
         if host is None:
             return jsonify({'message' : 'host is required in the query'}), 400
 
-        result = adddomaindaddy(host,new_domain)
-        if result == 1:
-            return jsonify({'message': 'the domain not found please correct this in the env and rebuild the container'}) , 500
-        elif result == 2:
-            return jsonify({'message': 'rate limit of godaddy triggered please handle this asyncr'}) , 204
-        elif result == 3:
-            return jsonify({'message' : "godaddy error "}) , 500
+        
 
         new_domain = generate_unique_hash()
-        config_filename = new_domain + '.conf'
-        with open(path_to_template, 'r') as template_file:
+        result = adddomaindaddy(new_domain)
+        if result is not None:
+            return jsonify({'message': result}), 400
+
+        with open(path_to_template_auto, 'r') as template_file:
             template_content = template_file.read()
-        template_content = template_content.replace(f'<$variable1>', str(new_domain))
-        template_content = template_content.replace(f'<$variable2>', str(''))
-        with open(os.path.join(config_output_path, config_filename), 'w') as output_file:
-            output_file.write(template_content)
-        subprocess.run(['nginx', '-s', 'reload'], check=True)  # Add check=True to raise an exception if the command fails
 
-
-
+        template_content = template_content.replace('<$variable1>', new_domain)
+        template_content = template_content.replace('<$domain>', DOMAINNAME)
+        template_content = template_content.replace('<$host>', HOST)
+        config_file_path = os.path.join(config_output_path, f"{new_domain}.{DOMAINNAME}.conf")
+        with open(config_file_path, 'w') as config_file:
+            config_file.write(template_content)
+        subprocess.run(['certbot', '--nginx', '--hsts' ,'-d' , f"{new_domain}.{DOMAINNAME}" ], check=True)  # Add check=True to raise an exception if the command fails
 
         return jsonify({'message': f'domain {new_domain} created successfully.'}), 200
     except Exception as e:
@@ -158,34 +157,74 @@ def create_domain_tenant_based():
         return jsonify({'message': 'method Not Allowed'}), 405
     tenant_id = request.form.get('tenant_id')
     subdomain = request.form.get('subdomain')
-    host = request.form.get('host')
+    sslkey_b64 = request.form.get('sslkey')
+    sslcertificat_b64 = request.form.get('sslcertificat')
     try:
-        if not tenant_id or not subdomain or not host:
+        if (sslkey_b64 and not sslcertificat_b64) or (sslcertificat_b64 and not sslkey_b64):
+            return jsonify({'message': 'Both SSL key and certificate are required if one is provided'}), 400
+
+        sslkey = None
+        sslcertificat = None
+
+        if sslkey_b64 and sslcertificat_b64:
+            sslkey = base64.b64decode(sslkey_b64).decode('ascii')
+            sslcertificat = base64.b64decode(sslcertificat_b64).decode('ascii')
+            if "----- END PRIVATE KEY -----" not in sslkey:
+                return jsonify({'message': 'Invalid SSL key format'}), 400
+            
+            if "----- END CERTIFICATE -----" not in sslcertificat:
+                return jsonify({'message': 'Invalid SSL certificate format'}), 400
+
+        if not tenant_id or not subdomain:
             return jsonify({'message': 'both tenant-id and subdomain parameters are required'}), 400
     
-        resutl = adddomaindaddy(host,subdomain)
+        result = adddomaindaddy(subdomain)
 
-        if resutl == 1:
-            return jsonify({'message': 'the domain not found please correct this in the env and rebuild the container'}) , 500
-        elif resutl == 2:
-            return jsonify({'message': 'rate limit of godaddy triggered please handle this asyncr'}) , 204
-        elif resutl == 3:
-            return jsonify({'message' : "godaddy error "}) , 500
+        if result is not None:
+            return jsonify({'message': result}), 400
+
+        if not sslkey and not sslcertificat:
+            with open(path_to_template_auto,'r') as template_file:
+                template_content = template_file.read()
+            template_content = template_content.replace('<$variable1>',subdomain)
+            template_content = template_content.replace('<$domain>',DOMAINNAME)
+            template_content = template_content.replace('<$host>',HOST)
+            config_file_path = os.path.join(config_output_path, f"{subdomain}.{DOMAINNAME}.conf")
+            with open(config_file_path, "w") as config_file:
+                config_file.write(template_content)
+            subprocess.run(['certbot', '--nginx', '--hsts' ,'-d' , f"{subdomain}.{DOMAINNAME}" ], check=True)
+            return jsonify({'message': f'Domain {subdomain} created successfully'}), 200
+                
 
 
-        config_filename = subdomain + '.conf'
+        # manual add cert defined by the user 
+        if sslkey and sslcertificat:
+            new_dir_name = subdomain + '.' + DOMAINNAME
+            new_dir_path = os.path.join('/letsencrypt/certs/live', new_dir_name)
+            os.makedirs(new_dir_path, exist_ok=True)
+            sslkey_path = os.path.join(new_dir_path, 'privkey.pem')
+            sslcertificat_path = os.path.join(new_dir_path, 'cert.pem')
 
-        replacements = { 'variable1': subdomain , 'variable2': '' }  # Updated variable names
-        with open(path_to_template, 'r') as template_file:
-            template_content = template_file.read()
+            with open(sslkey_path, 'w') as sslkey_file:
+                sslkey_file.write(sslkey)
 
-        for placeholder, value in replacements.items():
-            template_content = template_content.replace(f'<${placeholder}>', str(value))
+            with open(sslcertificat_path, 'w') as sslcertificat_file:
+                sslcertificat_file.write(sslcertificat)
 
-        with open(os.path.join(config_output_path, config_filename), 'w') as output_file:
-            output_file.write(template_content)
-        subprocess.run(['nginx', '-s', 'reload'], check=True)
-        return jsonify({'message': f'Domain {subdomain} created successfully'}), 200
+            with open(path_to_template_manu, 'r') as template_file:
+                template_content = template_file.read()
+
+            template_content = template_content.replace('<$variable1>', subdomain)
+            template_content = template_content.replace('<$domain>', DOMAINNAME)
+            template_content = template_content.replace('<$dir>',new_dir_path)
+            template_content = template_content.replace('<$host>', HOST)   
+            config_file_path = os.path.join(config_output_path, f"{domains}.{DOMAINNAME}.conf")
+            with open(config_file_path, 'w') as config_file:
+                config_file.write(template_content)
+            subprocess.run(['nginx', '-s', 'reload'], check=True)
+            return jsonify({'message': f'Domain {subdomain} created successfully'}), 200
+
+        
     except FileNotFoundError as e:
         return jsonify({'message' : 'file not found '}), 500
     except Exception as e:
@@ -197,45 +236,36 @@ def create_domain_tenant_based():
 def remove_domain():
     if request.method != 'POST':
         return jsonify({'message': 'Method Not Allowed'}), 405
+    method = request.form.get('type')
     domain = request.form.get('domain')
+    
     try:
-        if not domain:
-            return jsonify({'message': 'Domain parameter is required'}), 400
-
+        if not domain or not method :
+            return jsonify({'error': 'domain and method parameter is required'}), 400
         
-        resutl = removedomaindaddy(domain)
+        result = removedomaindaddy(domain)
 
-        if resutl == 1:
-            return jsonify({'message' : 'record not found'}), 400
-        elif resutl == 2:
-            return jsonify({'message' : 'godaddy rate limit handle this async'}), 204
-        elif resutl  == 3:
-            return jsonify({'message' : 'goaddy error'}), 500
-
+        if result is not None:
+            return jsonify({'error': result}), 400
         
+        file = f"{domain}.{DOMAINNAME}.conf"
 
+        config_file_path = os.path.join(config_output_path, file )
+        os.remove(config_file_path)
 
-        # Delete the associated configuration file
-        config_filename = domain + '.conf'
-        config_file_path = os.path.join(config_output_path, config_filename)
-        if os.path.exists(config_file_path):
-            os.remove(config_file_path)
-            subprocess.run(['nginx', '-s', 'reload'], check=True)
+        if method == "auto":
+            subprocess(["certbot","revoke","--cert-name", f"{domain}.{DOMAINNAME}" ,"--delete-after-revoke"],check=True)
             return jsonify({'message': f'Domain {domain} removed successfully'}), 200
-        else:
-            return jsonify({'message': f'Domain {domain} not found.'}), 404
+        if method == "manu":
+            new_dir_path = os.path.join('/letsencrypt/certs/live', f"{domain}.{DOMAINNAME}")
+            os.removedirs(new_dir_path)
+            subprocess(["nginx","-s","reload"],check=True)
+            return jsonify({'message': f'Domain {domain} removed successfully'}), 200
+
+
+        
     except Exception as e:
-        return jsonify({'message': f'An error occurred while removing the domain: {str(e)}'}), 500
-
-
-
-# Endpoint 6: 404 handling 
-@app.route('/', defaults={'path': ''})
-@app.route('/<path:path>')
-def handle_undefined_routes(path):
-    # Return an error response for any undefined route (HTTP status code 404) , i think it better to return all the endpoint instead of Not Found
-    return jsonify({'error': 'Not Found , please dont not expose this to external , this is a internal service only if you see this message from the external fix it'}), 404
-
+        return jsonify({'error': f'An error occurred while removing the domain: {str(e)}'}), 500
 
 if __name__ == '__main__':
     if checkdomaingodaddy(DOMAINNAME) == False:
